requirements.txt
django~=3.2.5
djangorestframework~=3.12.4
python-scrapyd-api~=2.1.2
scrapy~=2.5.0
pandas~=0.25.3
numpy~=1.19.5
requests~=2.26.0
uritools~=3.0.2
psycopg2~=2.9.1
sqlalchemy~=1.3.18
pymysql~=1.0.2
itemadapter~=0.3.0
setuptools~=49.2.0
mysqlclient==2.0.3
scrapyd==1.2.1
google-search-results==2.4.0
scrapy-splash==0.7.2
django-celery-beat==2.2.1
celery==5.1.2
django-enum-choices==2.1.3
user-agent==0.1.9
anytree==2.8.0
beautifulsoup4==4.9.3
redis==3.5.3
